Unified component generator and report
